*Minor change to satisfy Coverity. *Fixed a bug with the 3-12-2016 hotfix that broke the ability to set the end of slide position. *Fixed a bug with the 3-12-2016 hotfix that broke the chord list generation for RS exports. *Improved the slide up/down toggle functions so that notes that already have a pitched end of slide position have it reset to 0, so if a down slide is turned into an up slide, it can't retain a then incorrect end of slide position. *Various other corrections to data type in terms of whether or not they are signed, where appropriate.  The changes in this commit remove 51 signed/unsigned comparison compiler warnings (-Wextra).

+++ src/beat.c
@@ -13,7 +13,7 @@ int eof_beat_stats_cached = 0;	//Tracks whether the cached statistics for the pr
 
 long eof_get_beat(EOF_SONG * sp, unsigned long pos)
 {
-	long i;
+	unsigned long i;
 
 	eof_log("eof_get_beat() entered", 2);
 
@@ -199,7 +199,7 @@ void eof_change_accurate_ts(EOF_SONG * sp, char function)
 	sp->beat[sp->beats - 1]->ppqn = sp->beat[sp->beats - 2]->ppqn;	//The last beat's tempo is the same as the previous beat's
 }
 
-int eof_beat_is_anchor(EOF_SONG * sp, int cbeat)
+int eof_beat_is_anchor(EOF_SONG * sp, unsigned long cbeat)
 {
 	eof_log("eof_beat_is_anchor() entered", 2);
 
@@ -328,10 +328,10 @@ void eof_realign_beats(EOF_SONG * sp, int cbeat)
 	}
 }
 
-void eof_recalculate_beats(EOF_SONG * sp, int cbeat)
+void eof_recalculate_beats(EOF_SONG * sp, unsigned long cbeat)
 {
-	int i;
-	int last_anchor = eof_find_previous_anchor(sp, cbeat);
+	unsigned long i;
+	unsigned long last_anchor = eof_find_previous_anchor(sp, cbeat);
 	long next_anchor = eof_find_next_anchor(sp, cbeat);
 	int beats = 0;
 	int count = 1;
